Harden soundboard and delayed callback lifetimes

+++ Source/EffectsBaseView.h
@@ -209,7 +209,7 @@ protected:
         slider.setScrollWheelEnabled(false);
         slider.valueFromTextFunction = [](const String& s) -> float { return Decibels::decibelsToGain(s.getFloatValue()); };
 
-        slider.textFromValueFunction = [valuePrefix](float v) -> String { return valuePrefix + Decibels::toString(Decibels::gainToDecibels(v), 1); };
+        slider.textFromValueFunction = [valuePrefix](double v) -> String { return valuePrefix + Decibels::toString(Decibels::gainToDecibels(v), 1); };
 
     #if JUCE_IOS
         //slider->setPopupDisplayEnabled(true, false, this);
